refactor: 검증 코드 및 오타 수정 varification 에서 verification으로 오타를 수정했습니다. checkStringIsNum 메서드에서 변환결과를 변수에 저장하지 않도록 수정했습니다.
